+++ tests/testthat/test-esp_getTiles.R
@@ -15,27 +15,49 @@ test_that("tiles online", {
   poly <- esp_get_ccaa("La Rioja")
   expect_error(esp_getTiles(poly, type = "FFF"))
 
+
   # Skip test as tiles sometimes are not available
   skip_on_cran()
   skip_if_offline()
 
 
+  save_png <- function(code, width = 400, height = 400) {
+    path <- tempfile(fileext = ".png")
+    png(path, width = width, height = height)
+    on.exit(dev.off())
+    terra::plotRGB(code)
+
+    path
+  }
   expect_s4_class(esp_getTiles(poly), "SpatRaster")
   expect_message(esp_getTiles(poly,
     zoom = 5, verbose = TRUE,
     update_cache = TRUE
   ))
 
-  # Try with geom
+
+  s <- esp_getTiles(poly)
+
+
+  # With sfc
   geom <- sf::st_geometry(poly)
+
+  # Convert from bbox
+
+  bbox <- sf::st_bbox(poly)
+  expect_error(esp_getTiles(bbox))
+  expect_silent(esp_getTiles(sf::st_as_sfc(bbox)))
+
+  frombbox <- esp_getTiles(sf::st_as_sfc(bbox))
+
+  expect_s3_class(geom, "sfc")
 
   expect_silent(esp_getTiles(geom))
 
+  sfc <- esp_getTiles(geom)
 
   # From cache
   expect_message(esp_getTiles(poly, zoom = 5, verbose = TRUE))
-
-  expect_message(esp_getTiles(sf::st_geometry(poly), verbose = TRUE))
   expect_message(esp_getTiles(poly, verbose = TRUE))
   expect_message(esp_getTiles(
     poly,
@@ -47,12 +69,16 @@ test_that("tiles online", {
   # Single point
   point <- esp_get_ccaa("Madrid")
   point <- sf::st_transform(point, 3857)
-  point <- sf::st_sample(point, 1)
 
   expect_message(esp_getTiles(point,
     type = "RedTransporte.Carreteras",
     verbose = TRUE
   ))
+
+  p <- esp_getTiles(point,
+    type = "RedTransporte.Carreteras"
+  )
+
   expect_message(esp_getTiles(poly,
     type = "RedTransporte.Carreteras",
     verbose = TRUE, mask = TRUE
@@ -68,7 +94,7 @@ test_that("tiles online", {
   jpeg <- provs[provs$value == "jpeg", ]
 
   expect_message(esp_getTiles(poly,
-    type = jpeg$provider,
+    type = as.character(jpeg$provider[1]),
     verbose = TRUE
   ))
 
@@ -77,4 +103,29 @@ test_that("tiles online", {
   )
 
   expect_s4_class(s, "SpatRaster")
+
+  # Check layers
+  n <- expect_silent(esp_getTiles(poly,
+    type = "RedTransporte.Carreteras"
+  ))
+
+
+  expect_equal(terra::nlyr(n), 4)
+
+  opaque <- expect_silent(esp_getTiles(poly,
+    type = "RedTransporte.Carreteras",
+    transparent = FALSE
+  ))
+
+  expect_equal(terra::nlyr(opaque), 3)
+
+  # Run only on windows
+  skip_on_os(c("mac", "linux", "solaris"))
+
+  expect_snapshot_file(save_png(opaque), "opaque.png")
+  expect_snapshot_file(save_png(n), "transp.png")
+  expect_snapshot_file(save_png(s), "silent.png")
+  expect_snapshot_file(save_png(p), "point.png")
+  expect_snapshot_file(save_png(sfc), "sfc.png")
+  expect_snapshot_file(save_png(frombbox), "frombbox.png")
 })